Perf code structure
